fix: min memory leak at the time

+++ project2/stack.h
@@ -19,7 +19,7 @@ extern struct FamilyTreeNode end_loop;
 struct istack {
     struct FamilyTreeNode *value;
     struct istack *next;
-    int depth;
+    unsigned long long depth;
 };
 
 struct stack {
